+++ .env
@@ -4,3 +4,11 @@ VITE_GITHUB_OAUTH_CLIENT_ID=
 # appearances
 VITE_APPEARANCES_SHOW_IMPORT_PROJECT=false
 VITE_APPEARANCES_SHOW_TUTORIALS=true
+
+# api prefix
+VITE_BASE_URL=/api
+# api target
+# TODO: hidden in production
+VITE_API_TARGET=
+# namespace
+VITE_API_NAMESPACE=
